+++ Source/Game/DataStorage/Structs/W_Records.cs
@@ -44,7 +44,7 @@ namespace Game.DataStorage
         public uint Id;
         public uint QuestFeedbackEffectID;
         public byte WhenToDisplay;
-        public byte TargetType;
+        public int TargetType;
         public int TargetAsset;
         public uint PlayerConditionID;
         public ushort CombatConditionID;
