+++ Source/Game/Network/Packets/LootPackets.cs
@@ -42,13 +42,13 @@ namespace Game.Network.Packets
         {
             _worldPacket.WritePackedGuid(Owner);
             _worldPacket.WritePackedGuid(LootObj);
-            _worldPacket.WriteUInt8(FailureReason);
+            _worldPacket.WriteUInt8((byte)FailureReason);
             _worldPacket.WriteUInt8(AcquireReason);
-            _worldPacket.WriteUInt8(LootMethod);
+            _worldPacket.WriteUInt8((byte)LootMethod);
             _worldPacket.WriteUInt8(Threshold);
             _worldPacket.WriteUInt32(Coins);
-            _worldPacket.WriteUInt32(Items.Count);
-            _worldPacket.WriteUInt32(Currencies.Count);
+            _worldPacket.WriteInt32(Items.Count);
+            _worldPacket.WriteInt32(Currencies.Count);
             _worldPacket.WriteBit(Acquired);
             _worldPacket.WriteBit(AELooting);
             _worldPacket.FlushBits();
@@ -250,8 +250,8 @@ namespace Game.Network.Packets
             _worldPacket.WritePackedGuid(LootObj);
             _worldPacket.WriteInt32(MapID);
             _worldPacket.WriteUInt32(RollTime);
-            _worldPacket.WriteUInt8(ValidRolls);
-            _worldPacket.WriteUInt8(Method);
+            _worldPacket.WriteUInt8((byte)ValidRolls);
+            _worldPacket.WriteUInt8((byte)Method);
             Item.Write(_worldPacket);
         }
 
@@ -272,7 +272,7 @@ namespace Game.Network.Packets
             _worldPacket.WritePackedGuid(LootObj);
             _worldPacket.WritePackedGuid(Player);
             _worldPacket.WriteInt32(Roll);
-            _worldPacket.WriteUInt8(RollType);
+            _worldPacket.WriteUInt8((byte)RollType);
             Item.Write(_worldPacket);
             _worldPacket.WriteBit(Autopassed);
             _worldPacket.FlushBits();
@@ -295,7 +295,7 @@ namespace Game.Network.Packets
             _worldPacket.WritePackedGuid(LootObj);
             _worldPacket.WritePackedGuid(Winner);
             _worldPacket.WriteInt32(Roll);
-            _worldPacket.WriteUInt8(RollType);
+            _worldPacket.WriteUInt8((byte)RollType);
             Item.Write(_worldPacket);
             _worldPacket.WriteBit(MainSpec);
             _worldPacket.FlushBits();
@@ -366,7 +366,7 @@ namespace Game.Network.Packets
         public override void Write()
         {
             _worldPacket.WritePackedGuid(LootObj);
-            _worldPacket.WriteUInt32(Players.Count);
+            _worldPacket.WriteInt32(Players.Count);
             Players.ForEach(guid => _worldPacket.WritePackedGuid(guid));
         }
