+++ Source/Game/Network/Packets/ReferAFriendPackets.cs
@@ -63,7 +63,7 @@ namespace Game.Network.Packets
 
         public override void Write()
         {
-            _worldPacket .WriteInt32(Reason);
+            _worldPacket .WriteInt32((int)Reason);
             // Client uses this string only if Reason == ERR_REFER_A_FRIEND_NOT_IN_GROUP || Reason == ERR_REFER_A_FRIEND_SUMMON_OFFLINE_S
             // but always reads it from packet
             _worldPacket.WriteBits(Str.GetByteCount(), 6);
